+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix exclusion of `DIRECT` payments in receiver and wallet stats queries. [#1139](https://github.com/stellar/stellar-disbursement-platform-backend/pull/1139)
+
 ## [6.5.0](https://github.com/stellar/stellar-disbursement-platform-backend/releases/tag/6.5.0) ([diff](https://github.com/stellar/stellar-disbursement-platform-backend/compare/6.4.0...6.5.0))
 
 ### Added